requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
pytz>=2023.3
